Added performance monitor
- Added performance monitor to main server form. Monitors CPU/RAM usage
of the server process and network UP/DOWN speeds of the main network
interface card (the one that's connected to the internet). More resource
monitoring tools can be implemented into the PerformanceMonitor class.
- Fixed null pointer exception when remote clipboard was used but
client's clipboard is empty.
- Modified client's main listener thread. Not finished, should be main
point of concern now.
- Started implementing client-side pinging. Client currently sends PING
command tags to check if data can still be sent to server.