Added client GUID, Fun Manager, Window Manager, and icon
- Added a client GUID function to SystemUtils that returns the first 8
characters of the generated GUID and saves it into the application
properties. This will need to be reworked so that the GUID is saved
elsewhere such as registry, so it can be the same for any client
version.
- Added fun manager, currently only has a playable piano using
Console.Beep()
- Added an icon to the main form which is supposed to resemble two eyes,
one horizontal and one vertical.
- Added a notification that pops up at the taskbars notify icon that
displays a new connection's details.
- Fixed clipboard retrieval for the 100th time. Should now never error
out. But return [[EMPTY]] if it does.
- Added a window manager that retrieves open windows (their handler,
title, executable path, and visibility type). The window manager also
allows
control of each returned window.
- Fixed control forms not showing up until data was retrieved
- Implemented window handling into SystemUtils
- Added some code regions to SystemUtils to organize.